// InventoryManagementUI/src/openapi/services/model/softwares.ts
/**
 * Seyir Füzeleri Sistem Mühendisliği Envanter Uygulaması API
 * NodeJS - ExpressJS API Docs / Fevzi TEMİZ
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Employees } from './employees';


export interface Softwares { 
    id: number;
    name: string;
    companyEmail: string;
    companyPhone: string;
    description: string;
    lastAgreementDate: string;
    nextAgreementDate: string;
    employee: Employees;
}